# src/docs/swagger.yaml
definitions:
...
  gen-concept-api_api_dto.CreatePropertyRequest:
    properties:
      categoryId:
        type: integer
      dataType:
        maxLength: 15
        type: string
      description:
        maxLength: 1000
        type: string
      icon:
        maxLength: 1000
        type: string
      name:
        maxLength: 50
        minLength: 3
        type: string
      unit:
        maxLength: 15
        type: string
    required:
    - categoryId
    - name
    type: object
...
  gen-concept-api_api_dto.PropertyCategoryResponse:
    properties:
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      properties:
        items:
          $ref: '#/definitions/gen-concept-api_api_dto.PropertyResponse'
        type: array
    type: object
  gen-concept-api_api_dto.PropertyResponse:
    properties:
      category:
        $ref: '#/definitions/gen-concept-api_api_dto.PropertyCategoryResponse'
      dataType:
        type: string
      description:
        type: string
      icon:
        type: string
      id:
        type: integer
      name:
        type: string
      unit:
        type: string
    type: object
...
  gen-concept-api_api_dto.UpdatePropertyRequest:
    properties:
      categoryId:
        type: integer
      dataType:
        maxLength: 15
        type: string
      description:
        maxLength: 1000
        type: string
      icon:
        maxLength: 1000
        type: string
      name:
        type: string
      unit:
        maxLength: 15
        type: string
    type: object
  gen-concept-api_api_helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      resultCode:
        $ref: '#/definitions/gen-concept-api_api_helper.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/gen-concept-api_api_validation.ValidationError'
        type: array
    type: object
  gen-concept-api_api_helper.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 40401
    - 42901
    - 42902
    - 50001
    - 50002
    type: integer
    x-enum-varnames:
    - Success
    - ValidationError
    - AuthError
    - ForbiddenError
    - NotFoundError
    - LimiterError
    - OtpLimiterError
    - CustomRecovery
    - InternalError
  gen-concept-api_api_validation.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
  gen-concept-api_domain_filter.Filter:
    properties:
      filterType:
        description: text number
        type: string
      from:
        type: string
      to:
        type: string
      type:
        description: contains notContains equals notEqual startsWith lessThan lessThanOrEqual
          greaterThan greaterThanOrEqual inRange endsWith
        type: string
    type: object
...
  gen-concept-api_domain_filter.PagedList-gen-concept-api_api_dto_PropertyResponse:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/gen-concept-api_api_dto.PropertyResponse'
        type: array
      pageNumber:
        type: integer
      pageSize:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
  gen-concept-api_domain_filter.PaginationInputWithFilter:
    properties:
      filter:
        additionalProperties:
          $ref: '#/definitions/gen-concept-api_domain_filter.Filter'
        type: object
      pageNumber:
        type: integer
      pageSize:
        type: integer
      sort:
        items:
          $ref: '#/definitions/gen-concept-api_domain_filter.Sort'
        type: array
    type: object
  gen-concept-api_domain_filter.Sort:
    properties:
      colId:
        type: string
      sort:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/properties/:
    post:
      consumes:
      - application/json
      description: Create a Property
      parameters:
      - description: Create a Property
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/gen-concept-api_api_dto.CreatePropertyRequest'
      responses:
        "201":
          description: Property response
          schema:
            allOf:
            - $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/gen-concept-api_api_dto.PropertyResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Create a Property
      tags:
      - Properties
  /v1/properties/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Property
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: response
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Delete a Property
      tags:
      - Properties
    get:
      consumes:
      - application/json
      description: Get a Property
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: Property response
          schema:
            allOf:
            - $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/gen-concept-api_api_dto.PropertyResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get a Property
      tags:
      - Properties
    put:
      consumes:
      - application/json
      description: Update a Property
      parameters:
      - description: Id
        in: path
        name: id
        required: true
        type: integer
      - description: Update a Property
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/gen-concept-api_api_dto.UpdatePropertyRequest'
      responses:
        "200":
          description: Property response
          schema:
            allOf:
            - $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/gen-concept-api_api_dto.PropertyResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
        "404":
          description: Not found
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Update a Property
      tags:
      - Properties
  /v1/properties/get-by-filter:
    post:
      consumes:
      - application/json
      description: Get Properties
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/gen-concept-api_domain_filter.PaginationInputWithFilter'
      responses:
        "200":
          description: Property response
          schema:
            allOf:
            - $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/gen-concept-api_domain_filter.PagedList-gen-concept-api_api_dto_PropertyResponse'
              type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/gen-concept-api_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get Properties
      tags:
      - Properties
...
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"